refactor(sdk): Move oidc module to authentication::oidc
