+++ crates/matrix-sdk/src/authentication/oidc/registrations.rs
