+++ crates/matrix-sdk/src/client/mod.rs
@@ -74,7 +74,7 @@ use url::Url;
 
 use self::futures::SendRequest;
 #[cfg(feature = "experimental-oidc")]
-use crate::oidc::Oidc;
+use crate::authentication::oidc::Oidc;
 use crate::{
     authentication::{
         matrix::MatrixAuth, AuthCtx, AuthData, ReloadSessionCallback, SaveSessionCallback,
