+++ crates/matrix-sdk/src/lib.rs
@@ -47,8 +47,6 @@ pub mod event_handler;
 mod http_client;
 pub mod media;
 pub mod notification_settings;
-#[cfg(feature = "experimental-oidc")]
-pub mod oidc;
 pub mod pusher;
 pub mod room;
 pub mod room_directory_search;
